renamed rmq modules to match struct names and added migration notes
