+++ src/rmq/mod.rs
@@ -1,7 +1,7 @@
 //! Range minimum query data structures. These data structures allow for the calculation of the index of the
 //! minimum element in a range of a static array in constant-time. The implementations are located
-//! in the [`binary_rmq`] and [`fast_rmq`] modules.
+//! in the [`sparse`] and [`small`] modules.
 
-pub mod fast_rmq;
+pub mod small;
 
-pub mod binary_rmq;
+pub mod sparse;
